WordCamp: Change timing of text domain load to allow usage of user locale
It turns out that `wp_get_current_user`, which is needed for `get_user_locale`
to work correctly, is not available when mu-plugin code runs. This defers the
text domain load until the `plugins_loaded` hook so that function is available,
which ensures that the `wordcamporg` text domain is loaded for the user locale
rather than the site locale.

Previous attempts to fix this issue, r5518 and r5520, failed because they didn't
take into account the `wp_get_current_user` dependency.