WordCamp Reports: Call `WordCamp_Admin::meta_keys` statically for front end use.
The `$wordcamp_admin` global is only available in wp-admin, but a future commit will introduce a public version of this report, so we need to be able to call `meta_keys()` on the front end as well. Initializing `WordCamp_Admin` on the front end would have undesirable side-effects, but `meta_keys()` is already static, so we can just switch to calling it that way.